requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
billiard==4.2.1
celery==5.5.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.8.0
Django==4.2.20
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
exceptiongroup==1.2.2
inflection==0.5.1
iniconfig==2.1.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kombu==5.5.3
mysqlclient==2.2.7
packaging==25.0
pillow==11.2.1
pluggy==1.5.0
prompt_toolkit==3.0.51
PyJWT==2.9.0
pytest==8.3.5
pytest-cov==6.1.1
pytest-django==4.11.1
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.2
rpds-py==0.24.0
six==1.17.0
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
